+++ src/components/Dialog/SettingDialog/SettingDialog.vue
@@ -484,7 +484,6 @@ import {
   buildSongTrackAudioFileNameFromRawData,
 } from "@/store/utility";
 import {
-  isProduction,
   SavingSetting,
   EngineSettingType,
   ExperimentalSettingType,
@@ -495,6 +494,7 @@ import {
 } from "@/type/preload";
 import { createLogger } from "@/domain/frontend/log";
 import { useRootMiscSetting } from "@/composables/useRootMiscSetting";
+import { isProduction } from "@/helpers/platform";
 
 type SamplingRateOption = EngineSettingType["outputSamplingRate"];
