+++ src/components/Talk/AudioDetail.vue
@@ -85,7 +85,8 @@ import { computed, nextTick, ref, watch } from "vue";
 import AccentPhrase from "./AccentPhrase.vue";
 import ToolTip from "@/components/ToolTip.vue";
 import { useStore } from "@/store";
-import { AudioKey, isMac } from "@/type/preload";
+import { AudioKey } from "@/type/preload";
+import { isMac } from "@/helpers/platform";
 import { EngineManifest } from "@/openapi/models";
 import { useShiftKey, useAltKey } from "@/composables/useModifierKey";
 import { useHotkeyManager } from "@/plugins/hotkeyPlugin";
