+++ src/helpers/path.ts
@@ -1,6 +1,6 @@
 import * as windowsPath from "@std/path/windows";
 import * as posixPath from "@std/path/posix";
-import { isWindows } from "@/type/preload";
+import { isWindows } from "./platform";
 
 const path = isWindows ? windowsPath : posixPath;
